+++ aptos-move/e2e-benchmark/Cargo.toml
@@ -12,11 +12,21 @@ repository = { workspace = true }
 rust-version = { workspace = true }
 
 [dependencies]
+anyhow = { workspace = true }
+aptos-cached-packages = { workspace = true }
+aptos-crypto = { workspace = true }
+aptos-gas-algebra = { workspace = true }
+aptos-gas-profiling = { workspace = true }
 aptos-language-e2e-tests = { workspace = true }
 aptos-logger = { workspace = true }
+aptos-sdk = { workspace = true }
 aptos-transaction-generator-lib = { workspace = true }
 aptos-transaction-workloads-lib = { workspace = true }
 aptos-types = { workspace = true }
+aptos-vm-environment = { workspace = true }
+async-trait = { workspace = true }
 clap = { workspace = true }
+e2e-move-tests = { path = "../e2e-move-tests" }
 rand = { workspace = true }
 serde_json = { workspace = true }
+tokio = { workspace = true }
