+++ aptos-move/e2e-benchmark/src/lib.rs
@@ -1,0 +1,4 @@
+// Copyright (c) Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod gas_profiling;
