+++ aptos-move/e2e-move-tests/Cargo.toml
@@ -24,9 +24,7 @@ aptos-language-e2e-tests = { workspace = true }
 aptos-package-builder = { workspace = true }
 aptos-rest-client = { workspace = true }
 aptos-sdk = { workspace = true }
-aptos-transaction-generator-lib = { workspace = true }
 aptos-transaction-simulation = { workspace = true }
-aptos-transaction-workloads-lib = { workspace = true }
 aptos-types = { workspace = true }
 aptos-vm = { workspace = true, features = ["testing"] }
 aptos-vm-environment = { workspace = true }
@@ -41,7 +39,6 @@ move-symbol-pool = { workspace = true }
 once_cell = { workspace = true }
 project-root = { workspace = true }
 proptest = { workspace = true }
-rand = { workspace = true }
 rstest = { workspace = true }
 serde = { workspace = true }
 sha3 = { workspace = true }
