+++ aptos-move/e2e-move-tests/src/harness.rs
@@ -388,6 +388,11 @@ impl<O: OutputLogger> MoveHarnessImpl<O> {
     /// Runs a transaction and return gas used.
     pub fn evaluate_gas(&mut self, account: &Account, payload: TransactionPayload) -> u64 {
         let txn = self.create_transaction_payload(account, payload);
+        self.evaluate_gas_signed(txn)
+    }
+
+    /// Runs a transaction and return gas used.
+    pub fn evaluate_gas_signed(&mut self, txn: SignedTransaction) -> u64 {
         let output = self.run_raw(txn);
         assert_success!(output.status().to_owned());
         output.gas_used()
@@ -400,6 +405,14 @@ impl<O: OutputLogger> MoveHarnessImpl<O> {
         payload: TransactionPayload,
     ) -> (TransactionGasLog, u64, Option<FeeStatement>) {
         let txn = self.create_transaction_payload(account, payload);
+        self.evaluate_gas_with_profiler_signed(txn)
+    }
+
+    /// Runs a transaction with the gas profiler.
+    pub fn evaluate_gas_with_profiler_signed(
+        &mut self,
+        txn: SignedTransaction,
+    ) -> (TransactionGasLog, u64, Option<FeeStatement>) {
         let (output, gas_log) = self
             .executor
             .execute_transaction_with_gas_profiler(txn, &AuxiliaryInfo::default())
