+++ crates/transaction-workloads-lib/src/lib.rs
@@ -9,3 +9,4 @@ mod prebuilt_packages;
 mod token_workflow;
 
 pub use move_workloads::{EntryPoints, LoopType, MapType, MonotonicCounterType, OrderBookState};
+pub use token_workflow::TokenWorkflowKind;
